Move `Payload` into `zenoh_utils`
